transactions overfører korrekt

+++ src/main/java/broker/Transaction.java
@@ -8,11 +8,11 @@ public class Transaction {
     private String seller;
     private String buyer;
     private String stockName;
-    private int price;
+    private double price;
     private int quantity;
     private Date date;
 
-    public Transaction(Order sellerOrder, Order buyerOrder, int price, int quantity) {
+    public Transaction(Order sellerOrder, Order buyerOrder, double price, int quantity) {
         this.sellerOrder = sellerOrder;
         this.buyerOrder = buyerOrder;
         this.price = price;
@@ -23,7 +23,7 @@ public class Transaction {
         stockName = sellerOrder.getStock();
     }
 
-    public Transaction(String seller, String buyer, String stockName, int price, int quantity) {
+    public Transaction(String seller, String buyer, String stockName, double price, int quantity) {
         this.seller = seller;
         this.buyer = buyer;
         this.stockName = stockName;
@@ -36,7 +36,7 @@ public class Transaction {
         seller = (String) arr[0];
         buyer = (String) arr[1];
         stockName = (String) arr[2];
-        price = (Integer) arr[3];
+        price = (Double) arr[3];
         quantity = (Integer) arr[4];
         date = new Date();
     }
@@ -53,7 +53,7 @@ public class Transaction {
         return stockName;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
